Add docker functionality

+++ app/modules/web/service/auth/logout.go
@@ -1,16 +1,25 @@
 package auth
 
 import (
+	"fmt"
+
 	"github.com/BuildWithYou/fetroshop-api/app/domain/customer_accesses"
+	"github.com/BuildWithYou/fetroshop-api/app/helper/constant"
 	"github.com/BuildWithYou/fetroshop-api/app/helper/errorhelper"
 	"github.com/BuildWithYou/fetroshop-api/app/helper/gormhelper"
 	"github.com/BuildWithYou/fetroshop-api/app/helper/jwt"
+	"github.com/BuildWithYou/fetroshop-api/app/helper/validatorhelper"
 	"github.com/BuildWithYou/fetroshop-api/app/model"
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/utils"
 )
 
 func (svc *AuthServiceImpl) Logout(ctx *fiber.Ctx) (*model.Response, error) {
+	if validatorhelper.IsNotNil(svc.Err) {
+		fmt.Print("Error: ", svc.Err.Error()) // #marked: logging
+		return nil, errorhelper.Error500(constant.ERROR_GENERAL)
+	}
+
 	customerID := jwt.GetCustomerID(ctx)
 	identifier := jwt.GetAccessIdentifier(ctx)
 	result := svc.CustomerAccessRepo.Delete(&customer_accesses.CustomerAccess{
